chore: move hash32 unit test under hash module
