--- tests/hash.rs
@@ -1,21 +1,0 @@
-use sosi::hash::hash32;
-
-#[test]
-fn test() {
-    let data1: [u8; 1] = [0x62];
-    let data2: [u8; 2] = [0xc3, 0x97];
-    let data3: [u8; 3] = [0xe2, 0x99, 0xa5];
-    let data4: [u8; 4] = [0xe1, 0x80, 0xb9, 0x32];
-    let data5: [u8; 48] = [
-        0x01, 0xc0, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x14, 0x00, 0x00, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x14, 0x00, 0x00,
-        0x00, 0x18, 0x28, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00,
-    ];
-    assert_eq!(hash32(&[], 0, 0xbc9f1d34), 0xbc9f1d34);
-    assert_eq!(hash32(&data1, 1, 0xbc9f1d34), 0xef1345c4);
-    assert_eq!(hash32(&data2, 2, 0xbc9f1d34), 0x5b663814);
-    assert_eq!(hash32(&data3, 3, 0xbc9f1d34), 0x323c078f);
-    assert_eq!(hash32(&data4, 4, 0xbc9f1d34), 0xed21633a);
-    assert_eq!(hash32(&data5, 48, 0x12345678), 0xf333dabb);
-}
